Populate compiler/xla/hlo/evaluator with relevant files from compiler/xla/service, update users
PiperOrigin-RevId: 483297403

+++ tensorflow/compiler/xla/hlo/evaluator/hlo_evaluator_typed_visitor_complex64.cc
@@ -13,9 +13,9 @@ See the License for the specific language governing permissions and
 limitations under the License.
 ==============================================================================*/
 
-#include "tensorflow/compiler/xla/service/hlo_evaluator.h"
-#include "tensorflow/compiler/xla/service/hlo_evaluator_typed_visitor.h"
+#include "tensorflow/compiler/xla/hlo/evaluator/hlo_evaluator.h"
+#include "tensorflow/compiler/xla/hlo/evaluator/hlo_evaluator_typed_visitor.h"
 
 namespace xla {
-template class HloEvaluatorTypedVisitor<int32_t>;
+template class HloEvaluatorTypedVisitor<complex64>;
 }  // namespace xla
